Use cork() in streamFile
In recent versions of µWebSockets, there is now a warning telling you to cork() responses before writing to them for better performance. This does exactly that, and removes the warning.

```Warning: uWS.HttpResponse writes must be made from within a corked callback. See documentation for uWS.HttpResponse.cork and consult the user manual.```